Add new methods for push notification registration
These new methods explicitly use the UIUserNotificationType, which was
introduced in iOS 8, in their signatures. The old methods used either
NSIntegers or the deprecated UIRemoteNotificationType in their signatures,
but interally used these values as the former type if on iOS 8 or later.
The SDK was getting away with this because the underlying bit union types
are the same, so the legacy methods curry to our new methods.