Change argument type for legacy push notification method
Technically this is a "breaking change" TM, but it should have almost
no impact on Objectiv-C users, since the bitmask is
just a typedef for an integery. Swift users using this single
method would now have to call `rawValue` on their bitmask type.
This is acceptable to avoid warnings being emitted from the framework,
and 2 of the 3 legacy methods were doing this already.
These methods are deprecated and not long for this world.