Move Cache Cleanining
Instead of cleaning the legacy cache when the developer sets
their API credentials, do it when the webservice object is
instanstiated. This ensures the cache cleaning will happen
even if the developer sets the credential properties directly.

Because multiple WebServices objects could exist in one codebase,
we'll ensure the code cleaning is only run once using
dispatch_once.