Properly Decode CMFileMetadata Objects
Special case this specific type, since it doesn't
have a __class__ value, we must recognize it and return
the expected class. This means subclasses of CMFileMetadata
will not reserialize as the subclass. Since no one has asked
for the ability to subclass the metadata object, we won't
worry about this use case for now.